refactor: extract product read operations into dedicated service
